feat: split the application into modules
