+++ src/app/commons/commons.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { LogoComponent } from './components/logo/logo.component';
+
+@NgModule({
+  declarations: [LogoComponent],
+  imports: [CommonModule],
+  exports: [LogoComponent],
+})
+export class CommonsModule {}
