+++ src/app/commons/components/logo/logo.component.spec.ts
