+++ src/app/commons/components/logo/logo.component.ts
