+++ src/app/markup/footer/footer.component.scss
