+++ src/app/pages/courses-page/actions/actions.component.ts
