+++ src/app/pages/courses-page/courses-list/course-card/course-actions/course-actions.component.scss
