+++ src/app/pages/courses-page/courses-list/course-card/course-card.component.html
