+++ src/app/pages/courses-page/courses-list/course-card/course-card.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { ICourse } from 'src/app/interfaces/course.interface';
+import { ICourse } from 'src/app/pages/interfaces/course.interface';
 
 @Component({
   selector: 'app-course-card',
