+++ src/app/pages/courses-page/courses-list/course-card/course-tools/course-tools.component.scss
