+++ src/app/pages/courses-page/courses-list/course-card/course-tools/pipes/duration.pipe.spec.ts
