+++ src/app/pages/courses-page/courses-list/courses-list.component.scss
