+++ src/app/pages/courses-page/courses-list/courses-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ICourse } from 'src/app/interfaces/course.interface';
+import { ICourse } from 'src/app/pages/interfaces/course.interface';
 
 @Component({
   selector: 'app-courses-list',
