+++ src/app/pages/courses-page/courses-page.component.scss
