+++ src/app/pages/courses-page/filters/filters.component.scss
