+++ src/app/pages/courses-page/filters/search/search.component.html
