+++ src/app/pages/interfaces/course.interface.ts
